requirements.txt
feedparser==6.0.12
fpdf2==2.8.7
matplotlib==3.10.0
nltk==3.9.4
numpy==2.2.6
pandas==2.2.3
requests==2.32.3
seaborn==0.13.2
torch==2.11.0
transformers==5.5.0
vaderSentiment==3.3.2
wordcloud==1.9.6
yfinance==1.2.1
python-dotenv==1.1.0
